bitbake-user-manual: Removed pn-buildlist, package-depends.dot and pn-depends.dot
A long time ago when we switched to task basked execution we added
task-depends.dot and generated package-depends.dot and pn-depends.dot
for compatibility as best we could. The problem is they contain
partial data about the taskgraph, its incomplete and tends to confuse
users. Richard proposed we remove the compatibilty outputs and
just generate the one which contains definitive data.

I removed them from the manual where they appeared in the
"Generating Dependency Graphs" section.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>